Promiscuous Mode
What happens in normal mode, when a packet is received by a network card, it verifies that it belongs to itself. If not, it drops the packet normally, but in the promiscuous mode is used to accept all the packets that flow through the network card.
Promiscuous mode is used to monitor(sniff) network traffic.